rename in points layer
